+++ wazp/callbacks/dashboard.py
@@ -1,0 +1,78 @@
+import utils
+from dash import Input, Output, State, dash_table
+
+
+def get_callbacks(app):
+    """Return all callback functions for the dashboard tab.
+
+    Parameters
+    ----------
+    app : dash.Dash
+        Dash app object for which these callbacks are defined
+    """
+
+    @app.callback(
+        Output("table-container", "children"),
+        Input("table-container", "children"),
+        State("session-storage", "data"),
+    )
+    def create_input_data_table(
+        table_container_children: list,
+        app_storage: dict,
+    ):
+        """Create table to select videos to include in plots.
+
+        Parameters
+        ----------
+        table_container_children : list
+            list of html elements to pass to the table container
+        app_storage : dict
+            data held in temporary memory storage,
+            accessible to all tabs in the app
+
+        Returns
+        -------
+        table_container_children
+            list of html elements to pass to the table container
+        """
+
+        # Create table of videos with metadata with checkboxes
+        if not table_container_children:
+
+            # videos list as df
+            df_metadata = utils.df_from_metadata_yaml_files(
+                app_storage["config"]["videos_dir_path"],
+                app_storage["metadata_fields"],
+            )
+            df_metadata = df_metadata[
+                [app_storage["config"]["metadata_key_field_str"]]
+            ]
+
+            # table component
+            table_container_children = [
+                dash_table.DataTable(
+                    id="input-data-table",
+                    data=df_metadata.to_dict("records"),
+                    selected_rows=[],
+                    row_selectable="multi",
+                    fixed_rows={"headers": True},
+                    page_size=4,
+                    page_action="native",
+                    sort_action="native",
+                    sort_mode="single",
+                    style_table={
+                        "height": "200px",
+                        "maxHeight": "200px",
+                        "width": "100%",
+                        "maxWidth": "100%",
+                        "overflowY": "scroll",
+                        "overflowX": "scroll",
+                    },
+                    style_cell={  # refers to all cells (the whole table)
+                        "textAlign": "left",
+                        "padding": 7,
+                        "fontFamily": "Helvetica",
+                    },
+                )
+            ]
+        return table_container_children
